requirements.txt
anthropic>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
flask>=3.0.0
resend>=0.8.0
python-dotenv>=1.0.0
lxml>=5.1.0
